+++ index.html
@@ -6,7 +6,7 @@
     <script src="https://cdn.jsdelivr.net/npm/semantic-ui@2.5.0/dist/semantic.min.js"></script>
     <!-- <script src="./light-client-js/dist/index.js"></script> -->
     <script type="module">
-        import * as CkbLightClient from "./light-client-js/dist/index.js";
+        import * as CkbLightClient from "./wasm/light-client-js/dist/index.js";
         const LightClient = CkbLightClient.LightClient;
         console.log(LightClient);
         console.log(CkbLightClient);
