+++ light-client-lib/Cargo.toml
@@ -56,7 +56,7 @@ web-time = "1.1.0"
 wasm-bindgen = "0.2.63"
 idb = "0.6"
 serde-wasm-bindgen = "0.6.5"
-light-client-db-common = { path = "../light-client-db-common" }
+light-client-db-common = { path = "../wasm/light-client-db-common" }
 web-sys = "0.3.72"
 console_error_panic_hook = { version = "0.1.7" }
 serde_json = "1.0.134"
