+++ package.json
@@ -2,8 +2,8 @@
     "private": true,
     "name": "ckb-light-client",
     "workspaces": [
-        "./light-client-db-worker",
+        "./wasm/light-client-db-worker",
         "./light-client-wasm",
-        "./light-client-js"
+        "./wasm/light-client-js"
     ]
 }
