+++ wasm/light-client-db-common/src/tests.rs
