+++ wasm/light-client-db-worker/.gitignore
