+++ wasm/light-client-db-worker/Cargo.toml
