+++ wasm/light-client-db-worker/index.ts
