+++ wasm/light-client-db-worker/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "light-client-db-worker",
+    "name": "ckb-light-client-db-worker",
     "scripts": {
         "build-debug": "webpack --mode=development",
         "build": "webpack --mode=production",
