+++ wasm/light-client-db-worker/src/db.rs
