+++ wasm/light-client-db-worker/src/lib.rs
