+++ wasm/light-client-db-worker/src/util.rs
