+++ wasm/light-client-db-worker/webpack.config.js
