+++ wasm/light-client-js/.gitignore
