+++ wasm/light-client-js/esbuild.config.mjs
