+++ wasm/light-client-js/jest.config.js
