+++ wasm/light-client-js/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "light-client-js",
+  "name": "ckb-light-client-js",
   "version": "1.0.0",
   "main": "dist/index.js",
   "license": "MIT",
@@ -11,13 +11,13 @@
     "esbuild-plugin-inline-worker": "^0.1.1",
     "esbuild-plugin-polyfill-node": "^0.3.0",
     "jest": "^29.7.0",
-    "ts-jest": "^29.2.5"
+    "ts-jest": "^29.2.5",
+    "ckb-light-client-db-worker": "*",
+    "ckb-light-client-wasm": "*"
   },
   "dependencies": {
     "@ckb-ccc/core": "0.1.0-alpha.6",
     "async-mutex": "^0.5.0",
-    "light-client-db-worker": "file:../light-client-db-worker",
-    "light-client-wasm": "file:../light-client-wasm",
     "stream-browserify": "^3.0.0"
   },
   "scripts": {
