+++ wasm/light-client-js/src/db.worker.ts
@@ -1,5 +1,5 @@
 import { DbWorkerInitializeOptions } from "./types";
-import wasmModule from "light-client-db-worker";
+import wasmModule from "ckb-light-client-db-worker";
 onerror = event => {
     console.error(event);
 }
