+++ wasm/light-client-js/src/types.ts
