+++ wasm/light-client-js/src/worker.d.ts
