+++ wasm/light-client-js/tsconfig.json
